Refactor initialization process
The purpose of this change is to allow SotaUptaneClient to operate offline
(before provisioning has completed on-line). At the moment SotaUptaneClient
still requires provisioning to complete (see SotaUptaneClient::initialize()),
but this refactor is a step along the path outlined in
https://github.com/uptane/aktualizr/issues/8
