+++ src/libaktualizr/primary/reregistration_test.cc
@@ -4,7 +4,6 @@
 #include <vector>
 
 #include "httpfake.h"
-#include "libaktualizr/aktualizr.h"
 #include "libaktualizr/config.h"
 #include "uptane_test_common.h"
 #include "virtualsecondary.h"
