+++ src/libaktualizr/uptane/CMakeLists.txt
@@ -35,10 +35,14 @@ else(BUILD_OSTREE AND SOTA_PACKED_CREDENTIALS)
 endif(BUILD_OSTREE AND SOTA_PACKED_CREDENTIALS)
 
 
-add_aktualizr_test(NAME uptane SOURCES uptane_test.cc PROJECT_WORKING_DIRECTORY LIBRARIES uptane_generator_lib)
-set_property(SOURCE uptane_test.cc PROPERTY COMPILE_DEFINITIONS TEST_PKCS11_MODULE_PATH="${TEST_PKCS11_MODULE_PATH}")
+add_aktualizr_test(NAME uptane
+                   SOURCES uptane_test.cc
+                   PROJECT_WORKING_DIRECTORY
+                   LIBRARIES uptane_generator_lib provisioner_test_utils virtual_secondary)
+set_property(SOURCE uptane_test.cc
+             PROPERTY COMPILE_DEFINITIONS
+             TEST_PKCS11_MODULE_PATH="${TEST_PKCS11_MODULE_PATH}")
 set_tests_properties(test_uptane PROPERTIES LABELS "crypto")
-target_link_libraries(t_uptane virtual_secondary)
 
 add_aktualizr_test(NAME uptane_delegation SOURCES uptane_delegation_test.cc PROJECT_WORKING_DIRECTORY
                    ARGS "$<TARGET_FILE:uptane-generator>" LIBRARIES uptane_generator_lib)
